refactor: 2nd stage of 2 phase commit to rename files
