+++ src/app/numenera/model/Ability.model.ts
