+++ src/app/numenera/model/Type.model.ts
